split up schedule.cs
to make obsoletes easier to remove in the future
