+++ src/NServiceBus.Core/NServiceBus.Core.csproj
@@ -160,6 +160,7 @@
     <Compile Include="Persistence\Msmq\MsmqPersistence.cs" />
     <Compile Include="Persistence\Msmq\SubscriptionStorage\MsmqSubscriptionPersistence.cs" />
     <Compile Include="Pipeline\PipelineModifications.cs" />
+    <Compile Include="Schedule_Obsolete.cs" />
     <Compile Include="Scheduling\Scheduler.cs" />
     <Compile Include="Serialization\SerializationConfigExtensions.cs" />
     <Compile Include="Serialization\SerializationConfiguration.cs" />
